CTD_Gen_AI_Alloydb/SCR/files/GenAI-for-Alloydb/requirements.txt
Flask
asyncio
nest_asyncio
cloud-sql-python-connector==0.9.3
google-cloud-aiplatform
langchain
pgvector
nest-asyncio
asyncpg
transformers
ipython
markdown
google-cloud-secret-manager
google-cloud-alloydb
google-cloud-alloydb-connector
SQLAlchemy
pg8000
pandas